@@ -1,17 +1,17 @@
 module github.com/bilal-bhatti/zipline
 
-go 1.13
+go 1.15
 
 require (
 	github.com/fatih/structtag v1.2.0
 	github.com/getkin/kin-openapi v0.75.0
-	github.com/go-chi/chi v4.0.3+incompatible
-	github.com/go-openapi/spec v0.19.6
+	github.com/go-chi/chi/v5 v5.0.4
+	github.com/go-openapi/spec v0.20.3
 	github.com/go-toolsmith/astcopy v1.0.0
-	github.com/google/subcommands v1.0.1
-	github.com/google/wire v0.4.0
-	github.com/iancoleman/strcase v0.0.0-20191112232945-16388991a334
+	github.com/google/subcommands v1.2.0
+	github.com/google/wire v0.5.0
+	github.com/iancoleman/strcase v0.2.0
 	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.5.1
-	golang.org/x/tools v0.0.0-20200203222849-174f5c63c9f5
+	github.com/stretchr/testify v1.7.0
+	golang.org/x/tools v0.1.6
 )
